refactor: dedupe previewScreenshotTest task registration
CrabshellFeaturePlugin and app/build.gradle.kts each had a near-
identical block wiring up the previewScreenshotTest task (jvmTest
exclusion, task registration, output-dir cleanup). Extract it into a
shared registerPreviewScreenshotTestTask() extension function in
build-logic so both call sites stay in sync.
